Fix pkg structure
