+++ app/src/main/java/by/alexandr7035/banking/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import by.alexandr7035.banking.ui.core.AppContainerScreen
+import by.alexandr7035.banking.ui.app_host.AppContainerScreen
 import by.alexandr7035.banking.ui.theme.BankingAppTheme
 
 class MainActivity : ComponentActivity() {
