+++ app/src/main/java/by/alexandr7035/banking/ui/app_host/AppIntent.kt
@@ -1,4 +1,4 @@
-package by.alexandr7035.banking.ui.core
+package by.alexandr7035.banking.ui.app_host
 
 sealed class AppIntent {
     object EnterApp: AppIntent()
