+++ app/src/main/java/by/alexandr7035/banking/ui/app_host/navigation/model/NavResult.kt
@@ -1,0 +1,5 @@
+package by.alexandr7035.banking.ui.app_host.navigation.model
+
+enum class NavResult {
+    SHOULD_LOGOUT
+}
