--- app/src/main/java/by/alexandr7035/banking/ui/core/AppLoadingScreen.kt
@@ -1,82 +1,0 @@
-package by.alexandr7035.banking.ui.core
-
-import android.graphics.fonts.FontFamily
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.capitalize
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.intl.Locale
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.toUpperCase
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import by.alexandr7035.banking.R
-import by.alexandr7035.banking.ui.components.DotsProgressIndicator
-import by.alexandr7035.banking.ui.components.ScreenPreview
-import by.alexandr7035.banking.ui.theme.primaryFontFamily
-
-@Composable
-fun AppLoadingScreen() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.primary),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Image(
-            painter = painterResource(
-                id = R.drawable.logo_variant
-            ),
-            contentDescription = "Logo",
-            colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.onPrimary),
-            modifier = Modifier
-                .size(160.dp)
-        )
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        Text(
-            text = stringResource(id = R.string.app_name).uppercase(),
-            style = TextStyle(
-                fontSize = 40.sp,
-                fontFamily = primaryFontFamily,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.onPrimary,
-                textAlign = TextAlign.Center,
-            )
-        )
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        DotsProgressIndicator(
-            circleSize = 18.dp,
-            travelDistance = 20.dp,
-            circleColor = MaterialTheme.colorScheme.onPrimary
-        )
-    }
-}
-
-@Composable
-@Preview
-fun AppLoadingScreen_Preview() {
-    ScreenPreview {
-        AppLoadingScreen()
-    }
-}
