--- app/src/main/java/by/alexandr7035/banking/ui/core/navigation/NavResult.kt
@@ -1,5 +1,0 @@
-package by.alexandr7035.banking.ui.core.navigation
-
-enum class NavResult {
-    SHOULD_LOGOUT
-}
